Move form components to form dir
