+++ web/src/components/routes/LudosRoutes.tsx
@@ -5,7 +5,7 @@ import { Navigate, Route, Routes } from 'react-router-dom'
 import { contentKey, newKey, feedbackKey, frontpageKey, ldKey, puhviKey, sukoKey, updateKey } from './routes'
 import { Frontpage } from '../frontpage/Frontpage'
 import { Exams } from '../exam/Exams'
-import { AssignmentForm } from '../exam/assignment/AssignmentForm'
+import { AssignmentForm } from '../exam/assignment/form/AssignmentForm'
 import { Assignment } from '../assignmentpage/Assignment'
 import { HeaderMobile } from '../header/HeaderMobile'
 import { IS_MOBILE_QUERY } from '../../constants'
